backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
langchain==0.0.310
openai>=1.6.1,<2.0.0
faiss-cpu==1.7.4
pyyaml==6.0.1
python-dotenv==1.0.0
tiktoken==0.5.1
numpy==1.24.3
pypdf==3.16.0
docx2txt==0.8
pytesseract==0.3.10
Pillow==10.0.1
psutil==5.9.6 